Move LocalesCommonMistakes to common folder
