+++ moe/best/athenaeum/app/middleware/locales/common/BUILD
@@ -1,0 +1,9 @@
+load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_library")
+
+kt_jvm_library(
+    name = "common",
+    srcs = glob(["**/*.kt"]),
+    visibility = [
+        "//visibility:public"
+    ]
+)
