+++ moe/best/athenaeum/app/udc/social/BUILD
@@ -16,7 +16,7 @@ kt_jvm_library(
     deps = [
         "//moe/best/athenaeum/app/command:commands",
         "//moe/best/athenaeum/app/middleware:middleware",
-        "//moe/best/athenaeum/app/middleware/locales:locales",
+        "//moe/best/athenaeum/app/middleware/locales/common:common",
         "//moe/best/athenaeum/app/startup:startup",
         "//moe/best/athenaeum/app/startup/loader:loader",
         "@maven//:io_ktor_ktor_http_jvm",
